Make extend a module as well
